requirements.txt
requests==2.32.3
django==5.1.7
djangorestframework==3.15.2
django-filter==24.3
python-dotenv==1.0.1
drf-spectacular==0.28.0
celery[redis]==5.4.0
django-celery-beat==2.7.0
